Use an InputStream for incoming data
Resolves #306

Incoming user data will no longer be buffered in a local
File. Instead, bytes will be streamed directly to their destination,
whether that's an RDF Graph or the backend BinaryService.

In the event that a user-supplied message digest does not match the
server-computed message digest, the BinaryService::purgeContent method
will be called after the completion of BinaryService::setContent.