Move incoming binary digest handling to BinaryService
Resolves #316

This moves the processing of incoming binary streams to the
BinaryService by adding an additional BinaryService::setContent method.
This method accepts an additional argument: `algorithm`, and responds
with `CompletableFuture<String>` where that response string is the
value of the computed digest. If the value of that digest is null, one
should take that to mean that the supplied algorithm was not supported.

If the digest value does not match the expected value, it is the
reponsibility of the client code (e.g. in the HTTP layer) to call
BinaryService::purgeContent. A runtime exception, on the other hand,
would indicate a server error.