Add configuration key for PUT containment
Resolves #trellis-ldp/trellis-cassandra/57

With this change, resource creation via PUT can be configured to be either
contained or uncontained. By default, the configuration is uncontained.

The integration tests have been updated to consider this case.

The webdav component requires that creation via PUT is configured as
'contained' rather than 'uncontained', and there are some efficiency
gains for configuring it as 'contained', but there is a fallback in case
the configuration is different.